feat: Improve error messages in connector execution #311  (#76)
Console errorrs are unclear and expose implementation details to no-code
/ low-code users.

This PR proposes a way to bubble up errors from components in a way that
the end-user error is separated from the internal error information
(stack, low-level details). The `sendReq` method in `pkg/openai` is
modified to return human-friendly errors.
https://github.com/instill-ai/pipeline-backend/pull/311 is needed to
test these changes.

~The [fault](https://github.com/Southclaws/fault) package is chosen to
carry out this task.~ While `fault` was chosen originally, some of the
features we were using were no longer needed after refactoring the code
to adapt to the use of `temporal` workflows to execute sync pipelines.

- A logger is injected in the `openai.Client` struct, so we no longer
need to pass the request variables as part of the error context.
- The client doesn't know the component ID and can't attach it to the
end-user message. However, because communication between temporal
processes prevents error unwrapping, we can only use temporal errors to
pass the end-user message.

Therefore, it was simpler to use [our own
implementation](https://github.com/instill-ai/x/pull/13) of error
messages, at least while the required functionality doesn't get too
sofisticated.

### 🗒️ Notes

- Some coverage was added due to the rapid development in the `main`
branches, which made manual testing a bit unreliable. Let me know if
this is a potential change preventer and if we should focus on coverage
later.
- Additionally, the `frankban/quicktest` package was added in the tests.
I'm a fan of the package but I can rework the tests with `require` in
order to align with the current tests.
- Many integration tests were removed as they were outdated. They aren't
being run now, if there's a reason to keep them I can revert that
change.

![CleanShot 2023-12-15 at 14 45
30](https://github.com/instill-ai/pipeline-backend/assets/3977183/da89cd77-b4b8-401e-aae8-506c55ef0290)


![CleanShot 2023-12-15 at 14 44
32](https://github.com/instill-ai/pipeline-backend/assets/3977183/5512bf25-6bd6-4f64-abdd-d61ee3f35e33)